Change packet id to u64

+++ src/server/client/notifications.rs
@@ -23,7 +23,7 @@ impl Client {
 		final_vec.push(PacketType::Notification as u8);
 		final_vec.extend(&(n_type as u16).to_le_bytes());
 		final_vec.extend(&(final_size as u16).to_le_bytes());
-		final_vec.extend(&0u32.to_le_bytes()); // packet_id doesn't matter for notifications
+		final_vec.extend(&0u64.to_le_bytes()); // packet_id doesn't matter for notifications
 		final_vec.extend(data);
 
 		final_vec
